switch to static builds
Get rid of setup.py
drop coverage
update tox to use static builds
update python version in docker container so we can test with django3.2
