+++ MANIFEST.in
@@ -1,4 +1,4 @@
-include README.rst
+include README.md
 include LICENSE.md
 include requirements-tests.txt
 recursive-include docs *
