+++ README.md
@@ -1,9 +1,7 @@
-.. -*-restructuredtext-*-
-
+# cg-django-uaa
 
 This is a cloud.gov UAA authentication backend for Django. It also
 includes a handy "fake cloud.gov" that makes it easy to log in
 as any user during development.
 
-For documentation, please visit `cg-django-uaa.readthedocs.io
-<http://cg-django-uaa.readthedocs.io/>`_.
+For documentation, please visit [cg-django-uaa.readthedocs.io](http://cg-django-uaa.readthedocs.io/)
