+++ requirements-dev.txt
@@ -1,5 +1,4 @@
 flake8==3.2.1
-coverage>=4.3.4
 Sphinx==1.5.2
 sphinx-autobuild==0.6.0
 sphinx-rtd-theme==0.1.9
